requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
onnxruntime>=1.16.0
Pillow>=10.0.0
requests>=2.31.0
numpy<2.0.0